refactor: prep for differentiation
provided promisified mocked client methods,
isolate event emitter
